Add Core Library Desugaring
