+++ app/src/main/java/dev/iamwami/app/quotinews/QuotiNewsApp.kt
@@ -1,7 +1,7 @@
-package dev.iamwami.app.quotinews.di
+package dev.iamwami.app.quotinews
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class QuotinewsApplication: Application()
+class QuotiNewsApp: Application()
