+++ app/src/main/java/dev/iamwami/app/quotinews/ui/home/HighLightedNews.kt
@@ -20,7 +20,7 @@ import coil.compose.AsyncImage
 import dev.iamwami.app.quotinews.R
 import dev.iamwami.app.quotinews.model.News
 import dev.iamwami.app.quotinews.ui.theme.QuotiNewsTheme
-import dev.iamwami.app.quotinews.ui.util.SampleNewsApiDataProvider
+import dev.iamwami.app.quotinews.ui.utils.SampleNewsApiDataProvider
 import dev.iamwami.app.quotinews.util.dateFormatter
 
 @Composable
