+++ app/src/main/java/dev/iamwami/app/quotinews/ui/home/HomeScreen.kt
@@ -1,8 +1,6 @@
 package dev.iamwami.app.quotinews.ui.home
 
-import android.os.Build
 import android.util.Log
-import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.*
@@ -12,13 +10,11 @@ import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import dev.iamwami.app.quotinews.model.NewsFeed
-import dev.iamwami.app.quotinews.ui.home.components.RelatedNews
 
 
 /**
  *Home feed screen displaying just the article feed
  * */
-@RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun HomeFeedScreenWithNewsList(
     newsFeed: NewsFeed,
@@ -62,7 +58,6 @@ fun HomeFeedScreenWithNewsList(
 
 
 @OptIn(ExperimentalMaterial3Api::class)
-@RequiresApi(Build.VERSION_CODES.O)
 @Composable
 fun NewsList(
     newsFeed: NewsFeed,
