+++ app/src/main/java/dev/iamwami/app/quotinews/ui/utils/QuotiNewsDestination.kt
@@ -1,4 +1,4 @@
-package dev.iamwami.app.quotinews.ui.util
+package dev.iamwami.app.quotinews.ui.utils
 
 interface QuotiNewsDestination {
     val route: String
