+++ app/src/main/java/dev/iamwami/app/quotinews/util/Formatter.kt
@@ -1,14 +1,10 @@
 package dev.iamwami.app.quotinews.util
 
-import android.os.Build
-import androidx.annotation.RequiresApi
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
-@RequiresApi(Build.VERSION_CODES.O)
 
 fun dateFormatter(date: String): String {
     val oldDate = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss'Z'")
     return LocalDate.parse(date, oldDate).toString()
-
 }
